+++ refactoring_summary.md
@@ -1,0 +1,56 @@
+# NotionEventWorker Refactoring Summary
+
+## Problem
+The `NotionEventWorker` module was violating the Single Responsibility Principle by handling too many responsibilities:
+1. Event reception and parsing
+2. Page content fetching and processing
+3. PII detection
+4. Page archiving
+5. Recursive child page processing
+
+This made the code difficult to maintain, test, and understand.
+
+## Solution
+We refactored the code by extracting distinct responsibilities into their own modules:
+
+### 1. `PIIDetector.Workers.Event.NotionEventWorker`
+- **Single Responsibility**: Event handling and delegation
+- **Key Functions**: 
+  - `perform/1`: Processes Oban job and delegates to appropriate handlers
+  - `process_by_event_type/3`: Routes events to the correct processor
+
+### 2. `PIIDetector.Platform.Notion.EventDataExtractor`
+- **Single Responsibility**: Extract important data from Notion event payloads
+- **Key Functions**:
+  - `get_page_id_from_event/1`: Extract page ID from various event formats
+  - `get_user_id_from_event/1`: Extract user ID from various event formats
+
+### 3. `PIIDetector.Platform.Notion.PageProcessor`
+- **Single Responsibility**: Process Notion pages for PII
+- **Key Functions**:
+  - `process_page/3`: Main entry point for page processing
+  - `process_page_content/4`: Process full page content
+  - `archive_page/1`: Handle archiving when PII is found
+
+### 4. `PIIDetector.Platform.Notion.PIIPatterns`
+- **Single Responsibility**: Quick detection of common PII patterns
+- **Key Functions**:
+  - `check_for_obvious_pii/1`: Perform regex-based PII detection
+  - `extract_page_title/1`: Helper to extract page titles
+
+## Benefits of This Refactoring
+
+1. **Improved Maintainability**: Each module has a clear, focused responsibility
+2. **Better Testability**: Smaller, focused modules are easier to test in isolation
+3. **Enhanced Readability**: Code is more organized and easier to understand
+4. **Easier to Extend**: Adding new features or modifying existing ones is simpler
+5. **Separation of Concerns**: API interaction, event handling, and business logic are properly separated
+
+## Testing
+All existing tests pass, confirming that the refactoring preserves the original functionality while improving the code structure.
+
+## Future Improvements
+Further opportunities for improvement:
+1. Create a `NotionContentFetcher` module to handle all API interactions
+2. Extract recursive processing logic to its own module
+3. Implement more specific error handling for each module
